+++ Cocoa/Plug-Ins/Template Editor/ElementDLNG.m
@@ -18,7 +18,7 @@
 	value = CFSwapInt32BigToHost(tmp);
 }
 
-- (unsigned int)sizeOnDisk
+- (NSUInteger)sizeOnDisk
 {
 	return SIZE_ON_DISK;
 }
